Supabase/requirements.txt
# MQTT
paho-mqtt==1.6.1

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Supabase + PostgreSQL
psycopg2-binary==2.9.9
supabase==2.3.4
python-dotenv==1.0.0

# Utilidades
python-dateutil==2.8.2
